refactor test: change tests to use snapbox instead of expect-test
