+++ src/battery-pack/Cargo.toml
@@ -44,6 +44,5 @@ anyhow.workspace = true
 [dev-dependencies]
 assert_cmd = "2"
 battery-pack = { path = ".", features = ["build", "cli"] }
-expect-test.workspace = true
 tempfile.workspace = true
 toml.workspace = true
