+++ src/battery-pack/bphelper-cli/Cargo.toml
@@ -33,5 +33,5 @@ syntect = { version = "5", default-features = false, features = ["default-themes
 two-face = "0.5"
 
 [dev-dependencies]
-expect-test.workspace = true
+snapbox = { version = "1.2.1", features = ["debug"] }
 tempfile = "3"
