+++ src/battery-pack/bphelper-cli/src/tui.rs
@@ -1,4 +1,6 @@
 //! Interactive TUI for battery-pack CLI.
+#[cfg(test)]
+mod tests;
 
 use crate::registry::{
     BatteryPackDetail, BatteryPackSummary, CrateSource, InstalledPack, fetch_battery_pack_detail,
@@ -2401,1658 +2403,4 @@ fn apply_add_changes(changes: &[AddChange]) -> Result<()> {
     }
 
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use bphelper_manifest::{BatteryPackSpec, CrateSpec, DepKind};
-    use expect_test::expect;
-    use std::collections::{BTreeMap, BTreeSet};
-
-    // ====================================================================
-    // Fixture helpers
-    // ====================================================================
-
-    /// Create a minimal CrateSpec with sensible defaults.
-    fn crate_spec(version: &str) -> CrateSpec {
-        CrateSpec {
-            version: version.to_string(),
-            features: BTreeSet::new(),
-            dep_kind: DepKind::Normal,
-            optional: false,
-        }
-    }
-
-    /// Create a CrateSpec with features.
-    fn crate_spec_with_features(version: &str, features: &[&str]) -> CrateSpec {
-        CrateSpec {
-            version: version.to_string(),
-            features: features.iter().map(|s| s.to_string()).collect(),
-            dep_kind: DepKind::Normal,
-            optional: false,
-        }
-    }
-
-    /// Create a BatteryPackSpec with the given crates and features.
-    fn make_spec(crates: &[(&str, CrateSpec)], features: &[(&str, &[&str])]) -> BatteryPackSpec {
-        BatteryPackSpec {
-            name: "test-battery-pack".to_string(),
-            version: "1.0.0".to_string(),
-            description: "A test battery pack".to_string(),
-            repository: None,
-            keywords: Vec::new(),
-            crates: crates
-                .iter()
-                .map(|(name, spec)| (name.to_string(), spec.clone()))
-                .collect(),
-            features: features
-                .iter()
-                .map(|(name, deps)| {
-                    (
-                        name.to_string(),
-                        deps.iter().map(|d| d.to_string()).collect(),
-                    )
-                })
-                .collect(),
-            hidden: BTreeSet::new(),
-            templates: BTreeMap::new(),
-        }
-    }
-
-    /// Create an InstalledPackState directly (bypasses I/O).
-    fn make_installed_pack(name: &str, entries: Vec<CrateEntry>) -> InstalledPackState {
-        InstalledPackState {
-            name: format!("{}-battery-pack", name),
-            short_name: name.to_string(),
-            version: "1.0.0".to_string(),
-            entries,
-            features: BTreeMap::new(),
-        }
-    }
-
-    /// Create a CrateEntry with the given state.
-    fn make_entry(name: &str, enabled: bool, originally_enabled: bool) -> CrateEntry {
-        CrateEntry {
-            name: name.to_string(),
-            version: "0.1.0".to_string(),
-            features: Vec::new(),
-            dep_kind: DepKind::Normal,
-            original_dep_kind: DepKind::Normal,
-            group: "default".to_string(),
-            enabled,
-            originally_enabled,
-        }
-    }
-
-    /// Create an InstalledState from packs.
-    fn make_installed(packs: Vec<InstalledPackState>) -> InstalledState {
-        InstalledState {
-            packs,
-            selected_index: 0,
-        }
-    }
-
-    /// Create a BatteryPackDetail for DetailScreen tests.
-    fn make_detail(crates: &[&str], templates: &[&str], examples: &[&str]) -> BatteryPackDetail {
-        BatteryPackDetail {
-            name: "test-battery-pack".to_string(),
-            short_name: "test".to_string(),
-            version: "1.0.0".to_string(),
-            description: "Test battery pack".to_string(),
-            repository: Some("https://github.com/test/test".to_string()),
-            owners: Vec::new(),
-            crates: crates.iter().map(|s| s.to_string()).collect(),
-            extends: Vec::new(),
-            templates: templates
-                .iter()
-                .map(|name| crate::registry::TemplateInfo {
-                    name: name.to_string(),
-                    path: format!("templates/{}", name),
-                    description: None,
-                    repo_path: None,
-                })
-                .collect(),
-            examples: examples
-                .iter()
-                .map(|name| crate::registry::ExampleInfo {
-                    name: name.to_string(),
-                    description: None,
-                    repo_path: None,
-                })
-                .collect(),
-        }
-    }
-
-    // ====================================================================
-    // Tier 1: State logic tests
-    // ====================================================================
-
-    // --- CrateEntry ---
-
-    /// [verify tui.installed.show-state]
-    #[test]
-    fn version_info_no_features() {
-        let entry = CrateEntry {
-            name: "serde".to_string(),
-            version: "1.0.0".to_string(),
-            features: Vec::new(),
-            dep_kind: DepKind::Normal,
-            original_dep_kind: DepKind::Normal,
-            group: "default".to_string(),
-            enabled: true,
-            originally_enabled: true,
-        };
-        assert_eq!(entry.version_info(), "(1.0.0)");
-    }
-
-    /// [verify tui.installed.show-state]
-    #[test]
-    fn version_info_with_features() {
-        let entry = CrateEntry {
-            name: "serde".to_string(),
-            version: "1.0.0".to_string(),
-            features: vec!["derive".to_string(), "std".to_string()],
-            dep_kind: DepKind::Normal,
-            original_dep_kind: DepKind::Normal,
-            group: "default".to_string(),
-            enabled: true,
-            originally_enabled: true,
-        };
-        assert_eq!(entry.version_info(), "(1.0.0, features: derive, std)");
-    }
-
-    /// [verify tui.installed.dep-kind]
-    /// [verify tui.installed.show-state]
-    #[test]
-    fn version_info_dev_dep() {
-        let entry = CrateEntry {
-            name: "insta".to_string(),
-            version: "1.0.0".to_string(),
-            features: Vec::new(),
-            dep_kind: DepKind::Dev,
-            original_dep_kind: DepKind::Dev,
-            group: "default".to_string(),
-            enabled: true,
-            originally_enabled: true,
-        };
-        assert_eq!(entry.version_info(), "(1.0.0, dev)");
-    }
-
-    /// [verify tui.installed.dep-kind]
-    /// [verify tui.installed.show-state]
-    #[test]
-    fn version_info_build_dep_with_features() {
-        let entry = CrateEntry {
-            name: "cc".to_string(),
-            version: "1.0.0".to_string(),
-            features: vec!["parallel".to_string()],
-            dep_kind: DepKind::Build,
-            original_dep_kind: DepKind::Build,
-            group: "default".to_string(),
-            enabled: true,
-            originally_enabled: true,
-        };
-        assert_eq!(entry.version_info(), "(1.0.0, build, features: parallel)");
-    }
-
-    // --- InstalledState::toggle_selected ---
-
-    /// [verify tui.installed.toggle-crate]
-    #[test]
-    fn toggle_selected_flips_entry() {
-        let mut state = make_installed(vec![make_installed_pack(
-            "web",
-            vec![
-                make_entry("axum", true, true),
-                make_entry("tower", true, true),
-                make_entry("reqwest", false, false),
-            ],
-        )]);
-
-        // Toggle first entry (axum: true -> false)
-        state.toggle_selected();
-        assert!(!state.packs[0].entries[0].enabled);
-        assert!(state.packs[0].entries[1].enabled); // tower unchanged
-
-        // Toggle again (axum: false -> true)
-        state.toggle_selected();
-        assert!(state.packs[0].entries[0].enabled);
-    }
-
-    /// [verify tui.installed.toggle-crate]
-    #[test]
-    fn toggle_selected_targets_correct_entry_across_packs() {
-        let mut state = make_installed(vec![
-            make_installed_pack(
-                "web",
-                vec![
-                    make_entry("axum", true, true),
-                    make_entry("tower", true, true),
-                ],
-            ),
-            make_installed_pack(
-                "db",
-                vec![
-                    make_entry("sqlx", true, true),
-                    make_entry("sea-orm", false, false),
-                ],
-            ),
-        ]);
-
-        // Move to index 2 (sqlx in second pack) and toggle
-        state.selected_index = 2;
-        state.toggle_selected();
-        assert!(!state.packs[1].entries[0].enabled); // sqlx toggled off
-        assert!(state.packs[0].entries[0].enabled); // axum unchanged
-        assert!(state.packs[0].entries[1].enabled); // tower unchanged
-    }
-
-    /// [verify tui.installed.features] (partial: verifies group isolation, not group-level toggling)
-    #[test]
-    fn toggle_only_affects_target_not_other_groups() {
-        let mut state = make_installed(vec![make_installed_pack(
-            "web",
-            vec![
-                CrateEntry {
-                    name: "axum".to_string(),
-                    version: "0.7.0".to_string(),
-                    features: Vec::new(),
-                    dep_kind: DepKind::Normal,
-                    original_dep_kind: DepKind::Normal,
-                    group: "server".to_string(),
-                    enabled: true,
-                    originally_enabled: true,
-                },
-                CrateEntry {
-                    name: "reqwest".to_string(),
-                    version: "0.12.0".to_string(),
-                    features: Vec::new(),
-                    dep_kind: DepKind::Normal,
-                    original_dep_kind: DepKind::Normal,
-                    group: "client".to_string(),
-                    enabled: true,
-                    originally_enabled: true,
-                },
-            ],
-        )]);
-
-        // Toggle axum (server group)
-        state.selected_index = 0;
-        state.toggle_selected();
-        assert!(!state.packs[0].entries[0].enabled); // axum off
-        assert!(state.packs[0].entries[1].enabled); // reqwest (client) unchanged
-    }
-
-    // --- toggle constraint: feature dependencies ---
-
-    /// [verify tui.installed.toggle-crate]
-    #[test]
-    fn toggle_off_prevented_when_required_by_other_feature() {
-        // "axum" is in group "server" but also listed in the "networking" feature.
-        // "reqwest" is in group "client" and also in "networking".
-        // With both enabled and "networking" feature defined, disabling either
-        // should be prevented because the other keeps "networking" active.
-        let mut pack = make_installed_pack(
-            "web",
-            vec![
-                CrateEntry {
-                    name: "axum".to_string(),
-                    version: "0.7.0".to_string(),
-                    features: Vec::new(),
-                    dep_kind: DepKind::Normal,
-                    original_dep_kind: DepKind::Normal,
-                    group: "server".to_string(),
-                    enabled: true,
-                    originally_enabled: true,
-                },
-                CrateEntry {
-                    name: "reqwest".to_string(),
-                    version: "0.12.0".to_string(),
-                    features: Vec::new(),
-                    dep_kind: DepKind::Normal,
-                    original_dep_kind: DepKind::Normal,
-                    group: "client".to_string(),
-                    enabled: true,
-                    originally_enabled: true,
-                },
-            ],
-        );
-        // Both crates are in the "networking" feature
-        pack.features.insert(
-            "networking".to_string(),
-            BTreeSet::from(["axum".to_string(), "reqwest".to_string()]),
-        );
-
-        let mut state = make_installed(vec![pack]);
-
-        // Try to toggle off axum — should be prevented (reqwest keeps "networking" active)
-        state.selected_index = 0;
-        state.toggle_selected();
-        assert!(state.packs[0].entries[0].enabled); // still enabled
-
-        // Try to toggle off reqwest — should also be prevented
-        state.selected_index = 1;
-        state.toggle_selected();
-        assert!(state.packs[0].entries[1].enabled); // still enabled
-    }
-
-    /// [verify tui.installed.toggle-crate]
-    #[test]
-    fn toggle_off_allowed_when_no_cross_feature_dependency() {
-        // Crate in its own group with no cross-feature memberships.
-        let mut state = make_installed(vec![make_installed_pack(
-            "web",
-            vec![make_entry("axum", true, true)],
-        )]);
-
-        state.toggle_selected();
-        assert!(!state.packs[0].entries[0].enabled); // toggled off successfully
-    }
-
-    /// [verify tui.installed.toggle-crate]
-    #[test]
-    fn toggle_on_always_allowed_even_with_features() {
-        // Enabling a crate is always allowed, regardless of feature constraints.
-        let mut pack = make_installed_pack(
-            "web",
-            vec![
-                CrateEntry {
-                    name: "axum".to_string(),
-                    version: "0.7.0".to_string(),
-                    features: Vec::new(),
-                    dep_kind: DepKind::Normal,
-                    original_dep_kind: DepKind::Normal,
-                    group: "server".to_string(),
-                    enabled: false,
-                    originally_enabled: false,
-                },
-                CrateEntry {
-                    name: "reqwest".to_string(),
-                    version: "0.12.0".to_string(),
-                    features: Vec::new(),
-                    dep_kind: DepKind::Normal,
-                    original_dep_kind: DepKind::Normal,
-                    group: "client".to_string(),
-                    enabled: true,
-                    originally_enabled: true,
-                },
-            ],
-        );
-        pack.features.insert(
-            "networking".to_string(),
-            BTreeSet::from(["axum".to_string(), "reqwest".to_string()]),
-        );
-
-        let mut state = make_installed(vec![pack]);
-
-        // Toggle axum ON — should always succeed
-        state.selected_index = 0;
-        state.toggle_selected();
-        assert!(state.packs[0].entries[0].enabled);
-    }
-
-    // --- dep_kind cycling ---
-
-    /// [verify tui.installed.dep-kind]
-    #[test]
-    fn cycle_dep_kind_cycles_through_all_variants() {
-        let mut state = make_installed(vec![make_installed_pack(
-            "web",
-            vec![make_entry("axum", true, true)],
-        )]);
-
-        assert_eq!(state.packs[0].entries[0].dep_kind, DepKind::Normal);
-
-        state.cycle_dep_kind();
-        assert_eq!(state.packs[0].entries[0].dep_kind, DepKind::Dev);
-
-        state.cycle_dep_kind();
-        assert_eq!(state.packs[0].entries[0].dep_kind, DepKind::Build);
-
-        state.cycle_dep_kind();
-        assert_eq!(state.packs[0].entries[0].dep_kind, DepKind::Normal);
-    }
-
-    /// [verify tui.installed.dep-kind]
-    #[test]
-    fn cycle_dep_kind_targets_selected_entry() {
-        let mut state = make_installed(vec![make_installed_pack(
-            "web",
-            vec![
-                make_entry("axum", true, true),
-                make_entry("tower", true, true),
-            ],
-        )]);
-
-        // Cycle second entry
-        state.selected_index = 1;
-        state.cycle_dep_kind();
-        assert_eq!(state.packs[0].entries[0].dep_kind, DepKind::Normal); // axum unchanged
-        assert_eq!(state.packs[0].entries[1].dep_kind, DepKind::Dev); // tower cycled
-    }
-
-    /// [verify tui.installed.dep-kind]
-    #[test]
-    fn dep_kind_change_detected_by_has_changes() {
-        let mut state = make_installed(vec![make_installed_pack(
-            "web",
-            vec![make_entry("axum", true, true)],
-        )]);
-
-        assert!(!state.has_changes());
-        state.cycle_dep_kind();
-        assert!(state.has_changes());
-
-        // Cycle back to original — no longer a change
-        state.cycle_dep_kind(); // Dev -> Build
-        state.cycle_dep_kind(); // Build -> Normal
-        assert!(!state.has_changes());
-    }
-
-    // --- InstalledState navigation ---
-
-    #[test]
-    fn installed_navigation_wraps() {
-        let mut state = make_installed(vec![make_installed_pack(
-            "web",
-            vec![
-                make_entry("a", true, true),
-                make_entry("b", true, true),
-                make_entry("c", true, true),
-            ],
-        )]);
-
-        assert_eq!(state.selected_index, 0);
-        state.select_prev(); // wrap to end
-        assert_eq!(state.selected_index, 2);
-        state.select_next(); // wrap to start
-        assert_eq!(state.selected_index, 0);
-    }
-
-    // --- InstalledState::has_changes / has_new_packs ---
-
-    /// [verify tui.nav.exit]
-    #[test]
-    fn has_changes_detects_toggled_entries() {
-        let state = make_installed(vec![make_installed_pack(
-            "web",
-            vec![
-                make_entry("axum", true, true),   // no change
-                make_entry("tower", false, true), // was on, now off
-            ],
-        )]);
-        assert!(state.has_changes());
-    }
-
-    #[test]
-    fn has_changes_false_when_unchanged() {
-        let state = make_installed(vec![make_installed_pack(
-            "web",
-            vec![
-                make_entry("axum", true, true),
-                make_entry("tower", false, false),
-            ],
-        )]);
-        assert!(!state.has_changes());
-    }
-
-    /// [verify tui.browse.add]
-    #[test]
-    fn has_new_packs_detects_added_from_browse() {
-        let state = make_installed(vec![make_installed_pack(
-            "web",
-            vec![
-                // All originally_enabled=false (came from Browse), some now enabled
-                make_entry("axum", true, false),
-                make_entry("tower", false, false),
-            ],
-        )]);
-        assert!(state.has_new_packs());
-    }
-
-    #[test]
-    fn has_new_packs_false_when_none_enabled() {
-        let state = make_installed(vec![make_installed_pack(
-            "web",
-            vec![
-                make_entry("axum", false, false),
-                make_entry("tower", false, false),
-            ],
-        )]);
-        assert!(!state.has_new_packs());
-    }
-
-    // --- InstalledState::collect_changes ---
-
-    /// [verify tui.nav.exit]
-    #[test]
-    fn collect_changes_new_pack() {
-        let state = make_installed(vec![make_installed_pack(
-            "web",
-            vec![
-                make_entry("axum", true, false),
-                make_entry("tower", true, false),
-                make_entry("reqwest", false, false), // not selected
-            ],
-        )]);
-
-        let changes = state.collect_changes();
-        expect![[r#"
-            [
-                AddPack {
-                    name: "web-battery-pack",
-                    crates: [
-                        CrateChange {
-                            name: "axum",
-                            dep_kind: Normal,
-                        },
-                        CrateChange {
-                            name: "tower",
-                            dep_kind: Normal,
-                        },
-                    ],
-                },
-            ]
-        "#]]
-        .assert_debug_eq(&changes);
-    }
-
-    /// [verify tui.nav.exit]
-    #[test]
-    fn collect_changes_update_existing_pack() {
-        let state = make_installed(vec![make_installed_pack(
-            "web",
-            vec![
-                make_entry("axum", true, true),     // unchanged
-                make_entry("tower", false, true),   // removed
-                make_entry("reqwest", true, false), // added
-            ],
-        )]);
-
-        let changes = state.collect_changes();
-        expect![[r#"
-            [
-                UpdatePack {
-                    name: "web-battery-pack",
-                    add_crates: [
-                        CrateChange {
-                            name: "reqwest",
-                            dep_kind: Normal,
-                        },
-                    ],
-                    remove_crates: [
-                        "tower",
-                    ],
-                },
-            ]
-        "#]]
-        .assert_debug_eq(&changes);
-    }
-
-    /// [verify tui.nav.exit]
-    #[test]
-    fn collect_changes_skips_unchanged_packs() {
-        let state = make_installed(vec![
-            make_installed_pack(
-                "web",
-                vec![
-                    make_entry("axum", true, true),
-                    make_entry("tower", true, true),
-                ],
-            ),
-            make_installed_pack(
-                "db",
-                vec![
-                    make_entry("sqlx", false, true), // changed
-                ],
-            ),
-        ]);
-
-        let changes = state.collect_changes();
-        expect![[r#"
-            [
-                UpdatePack {
-                    name: "db-battery-pack",
-                    add_crates: [],
-                    remove_crates: [
-                        "sqlx",
-                    ],
-                },
-            ]
-        "#]]
-        .assert_debug_eq(&changes);
-    }
-
-    // --- DetailScreen ---
-
-    /// [verify tui.browse.detail]
-    #[test]
-    fn detail_selectable_items_includes_all_sections() {
-        let detail = make_detail(
-            &["serde", "tokio"],    // 2 crates
-            &["basic", "advanced"], // 2 templates
-            &["hello-world"],       // 1 example
-        );
-        let screen = DetailScreen {
-            detail: Rc::new(detail),
-            selected_index: 0,
-            came_from_list: false,
-        };
-
-        let items: Vec<_> = screen.selectable_items().collect();
-        assert_eq!(screen.item_count(), items.len());
-
-        expect![[r#"
-            [
-                Crate(
-                    "serde",
-                ),
-                Crate(
-                    "tokio",
-                ),
-                Template {
-                    _path: "templates/basic",
-                    repo_path: None,
-                },
-                Template {
-                    _path: "templates/advanced",
-                    repo_path: None,
-                },
-                Example {
-                    _name: "hello-world",
-                    repo_path: None,
-                },
-                ActionOpenCratesIo,
-                ActionAddToProject,
-                ActionNewProject,
-            ]
-        "#]]
-        .assert_debug_eq(&items);
-    }
-
-    /// [verify tui.browse.detail]
-    #[test]
-    fn detail_navigation_wraps() {
-        let detail = make_detail(&["serde"], &[], &[]);
-        let mut screen = DetailScreen {
-            detail: Rc::new(detail),
-            selected_index: 0,
-            came_from_list: false,
-        };
-
-        // 1 crate + 3 actions = 4 items
-        assert_eq!(screen.item_count(), 4);
-
-        // Navigate to last item
-        screen.select_prev(); // wraps to 3
-        assert_eq!(screen.selected_index, 3);
-
-        // Navigate forward wraps to 0
-        screen.select_next();
-        assert_eq!(screen.selected_index, 0);
-    }
-
-    #[test]
-    fn detail_selected_item_returns_correct_item() {
-        let detail = make_detail(&["serde", "tokio"], &[], &[]);
-        let screen = DetailScreen {
-            detail: Rc::new(detail),
-            selected_index: 1, // tokio
-            came_from_list: false,
-        };
-
-        let item = screen.selected_item().unwrap();
-        assert!(matches!(item, DetailItem::Crate(n) if n == "tokio"));
-    }
-
-    // --- ExpandedPack ---
-
-    /// [verify tui.browse.add]
-    #[test]
-    fn expanded_pack_toggle_and_navigate() {
-        let mut expanded = ExpandedPack {
-            pack: make_installed_pack(
-                "web",
-                vec![
-                    make_entry("axum", true, false),
-                    make_entry("tower", true, false),
-                    make_entry("reqwest", false, false),
-                ],
-            ),
-            selected_index: 0,
-        };
-
-        // Toggle first entry off
-        expanded.toggle_selected();
-        assert!(!expanded.pack.entries[0].enabled);
-
-        // Navigate to reqwest (index 2)
-        expanded.select_next();
-        expanded.select_next();
-        assert_eq!(expanded.selected_index, 2);
-
-        // Toggle reqwest on
-        expanded.toggle_selected();
-        assert!(expanded.pack.entries[2].enabled);
-
-        // Wrap navigation
-        expanded.select_next(); // wraps to 0
-        assert_eq!(expanded.selected_index, 0);
-    }
-
-    // --- build_installed_state (integration with BatteryPackSpec) ---
-
-    /// [verify tui.installed.show-state]
-    #[test]
-    fn build_installed_state_from_spec() {
-        let spec = make_spec(
-            &[
-                ("serde", crate_spec("1.0.0")),
-                ("tokio", crate_spec_with_features("1.0.0", &["full"])),
-            ],
-            &[("default", &["serde", "tokio"])],
-        );
-
-        let installed_pack = crate::registry::InstalledPack {
-            name: "test-battery-pack".to_string(),
-            short_name: "test".to_string(),
-            version: "1.0.0".to_string(),
-            spec,
-            active_features: BTreeSet::from(["default".to_string()]),
-        };
-
-        let state = build_installed_state(vec![installed_pack]);
-        assert_eq!(state.packs.len(), 1);
-
-        let pack = &state.packs[0];
-        assert_eq!(pack.name, "test-battery-pack");
-        assert_eq!(pack.entries.len(), 2);
-
-        // Both should be enabled (in active features)
-        assert!(pack.entries.iter().all(|e| e.enabled));
-        assert!(pack.entries.iter().all(|e| e.originally_enabled));
-
-        // Check tokio has features
-        let tokio_entry = pack.entries.iter().find(|e| e.name == "tokio").unwrap();
-        assert_eq!(tokio_entry.features, vec!["full"]);
-    }
-
-    /// [verify tui.browse.add]
-    #[test]
-    fn build_expanded_pack_defaults_prechecked() {
-        let spec = make_spec(
-            &[
-                ("serde", crate_spec("1.0.0")),
-                ("tokio", crate_spec("1.0.0")),
-                ("tracing", crate_spec("0.1.0")),
-            ],
-            &[("default", &["serde", "tokio"])],
-        );
-
-        let summary = BatteryPackSummary {
-            name: "test-battery-pack".to_string(),
-            short_name: "test".to_string(),
-            version: "1.0.0".to_string(),
-            description: String::new(),
-        };
-
-        let expanded = build_expanded_pack(&summary, spec);
-
-        // All originally_enabled should be false (new pack)
-        assert!(expanded.pack.entries.iter().all(|e| !e.originally_enabled));
-
-        // Default crates should be pre-checked (enabled=true)
-        let serde = expanded
-            .pack
-            .entries
-            .iter()
-            .find(|e| e.name == "serde")
-            .unwrap();
-        assert!(serde.enabled, "default crate serde should be pre-checked");
-
-        let tokio = expanded
-            .pack
-            .entries
-            .iter()
-            .find(|e| e.name == "tokio")
-            .unwrap();
-        assert!(tokio.enabled, "default crate tokio should be pre-checked");
-
-        // Non-default crates should not be pre-checked
-        let tracing = expanded
-            .pack
-            .entries
-            .iter()
-            .find(|e| e.name == "tracing")
-            .unwrap();
-        assert!(
-            !tracing.enabled,
-            "non-default crate tracing should not be pre-checked"
-        );
-    }
-
-    /// Helper: create an App with a given screen (bypasses loading).
-    fn make_app(screen: Screen) -> App {
-        App {
-            source: CrateSource::Registry,
-            screen,
-            should_quit: false,
-            pending_action: None,
-        }
-    }
-
-    /// Helper: create an AddScreen with installed packs and empty browse.
-    fn make_add_screen(packs: Vec<InstalledPackState>) -> AddScreen {
-        AddScreen {
-            tab: AddTab::Installed,
-            installed: make_installed(packs),
-            browse: BrowseState {
-                items: Vec::new(),
-                list_state: ListState::default(),
-                search_input: String::new(),
-                searching: false,
-                expanded: None,
-            },
-            changes: None,
-        }
-    }
-
-    /// Create a BatteryPackSummary with sensible defaults.
-    fn make_summary(short_name: &str, version: &str, desc: &str) -> BatteryPackSummary {
-        BatteryPackSummary {
-            name: format!("{}-battery-pack", short_name),
-            short_name: short_name.to_string(),
-            version: version.to_string(),
-            description: desc.to_string(),
-        }
-    }
-
-    /// Extract the AddScreen from an App, panicking if it's a different screen.
-    fn unwrap_add_screen(app: &App) -> &AddScreen {
-        match &app.screen {
-            Screen::Add(state) => state,
-            _ => panic!("Expected Add screen"),
-        }
-    }
-
-    fn unwrap_list_screen(app: &App) -> &ListScreen {
-        match &app.screen {
-            Screen::List(state) => state,
-            _ => panic!("Expected List screen"),
-        }
-    }
-
-    fn unwrap_detail_screen(app: &App) -> &DetailScreen {
-        match &app.screen {
-            Screen::Detail(state) => state,
-            _ => panic!("Expected Detail screen"),
-        }
-    }
-
-    /// [verify tui.installed.show-state]
-    #[test]
-    fn build_installed_state_partial_features() {
-        let spec = make_spec(
-            &[
-                ("serde", crate_spec("1.0.0")),
-                ("tokio", crate_spec("1.0.0")),
-                ("tracing", crate_spec("0.1.0")),
-            ],
-            &[
-                ("default", &["serde"]),
-                ("async", &["tokio"]),
-                ("observability", &["tracing"]),
-            ],
-        );
-
-        let installed_pack = crate::registry::InstalledPack {
-            name: "test-battery-pack".to_string(),
-            short_name: "test".to_string(),
-            version: "1.0.0".to_string(),
-            spec,
-            // Only default + async active, not observability
-            active_features: BTreeSet::from(["default".to_string(), "async".to_string()]),
-        };
-
-        let state = build_installed_state(vec![installed_pack]);
-        let pack = &state.packs[0];
-
-        let serde = pack.entries.iter().find(|e| e.name == "serde").unwrap();
-        assert!(serde.enabled, "serde should be enabled (in default)");
-
-        let tokio = pack.entries.iter().find(|e| e.name == "tokio").unwrap();
-        assert!(tokio.enabled, "tokio should be enabled (in async)");
-
-        let tracing = pack.entries.iter().find(|e| e.name == "tracing").unwrap();
-        assert!(
-            !tracing.enabled,
-            "tracing should be disabled (observability not active)"
-        );
-    }
-
-    // ====================================================================
-    // Tier 2: Key handling tests
-    // ====================================================================
-
-    // --- List screen navigation ---
-
-    /// [verify tui.nav.keyboard]
-    #[test]
-    fn list_j_k_navigation() {
-        let mut list_state = ListState::default();
-        list_state.select(Some(0));
-        let mut app = make_app(Screen::List(ListScreen {
-            items: vec![
-                make_summary("a", "1.0.0", "Pack A"),
-                make_summary("b", "1.0.0", "Pack B"),
-            ],
-            list_state,
-            filter: None,
-        }));
-
-        app.handle_key(KeyCode::Char('j')); // down
-        assert_eq!(unwrap_list_screen(&app).list_state.selected(), Some(1));
-
-        app.handle_key(KeyCode::Char('k')); // back up
-        assert_eq!(unwrap_list_screen(&app).list_state.selected(), Some(0));
-    }
-
-    /// [verify tui.nav.keyboard]
-    #[test]
-    fn list_q_quits() {
-        let mut list_state = ListState::default();
-        list_state.select(Some(0));
-        let mut app = make_app(Screen::List(ListScreen {
-            items: vec![make_summary("a", "1.0.0", "")],
-            list_state,
-            filter: None,
-        }));
-
-        app.handle_key(KeyCode::Char('q'));
-        assert!(app.should_quit);
-    }
-
-    /// [verify tui.nav.keyboard]
-    #[test]
-    fn list_esc_quits() {
-        let mut list_state = ListState::default();
-        list_state.select(Some(0));
-        let mut app = make_app(Screen::List(ListScreen {
-            items: Vec::new(),
-            list_state,
-            filter: None,
-        }));
-
-        app.handle_key(KeyCode::Esc);
-        assert!(app.should_quit);
-    }
-
-    // --- Detail screen key handling ---
-
-    /// [verify tui.nav.keyboard]
-    #[test]
-    fn detail_tab_and_arrows_navigate() {
-        let detail = make_detail(&["serde", "tokio"], &[], &[]);
-        let mut app = make_app(Screen::Detail(DetailScreen {
-            detail: Rc::new(detail),
-            selected_index: 0,
-            came_from_list: false,
-        }));
-
-        // Tab moves forward
-        app.handle_key(KeyCode::Tab);
-        assert_eq!(unwrap_detail_screen(&app).selected_index, 1);
-
-        // Down arrow also moves forward
-        app.handle_key(KeyCode::Down);
-        assert_eq!(unwrap_detail_screen(&app).selected_index, 2);
-
-        // Up arrow moves back
-        app.handle_key(KeyCode::Up);
-        assert_eq!(unwrap_detail_screen(&app).selected_index, 1);
-    }
-
-    /// [verify tui.nav.keyboard]
-    #[test]
-    fn detail_esc_when_came_from_list_goes_back() {
-        let detail = make_detail(&["serde"], &[], &[]);
-        let mut app = make_app(Screen::Detail(DetailScreen {
-            detail: Rc::new(detail),
-            selected_index: 0,
-            came_from_list: true,
-        }));
-
-        // Esc with came_from_list transitions to Loading (process_loading runs in the
-        // main loop, not inline). The important thing: it didn't quit.
-        app.handle_key(KeyCode::Esc);
-        assert!(!app.should_quit);
-    }
-
-    /// [verify tui.nav.keyboard]
-    #[test]
-    fn detail_esc_when_not_from_list_quits() {
-        let detail = make_detail(&["serde"], &[], &[]);
-        let mut app = make_app(Screen::Detail(DetailScreen {
-            detail: Rc::new(detail),
-            selected_index: 0,
-            came_from_list: false,
-        }));
-
-        app.handle_key(KeyCode::Esc);
-        assert!(app.should_quit);
-    }
-
-    /// [verify tui.nav.keyboard]
-    #[test]
-    fn detail_q_quits() {
-        let detail = make_detail(&["serde"], &[], &[]);
-        let mut app = make_app(Screen::Detail(DetailScreen {
-            detail: Rc::new(detail),
-            selected_index: 0,
-            came_from_list: true, // even when came_from_list, q quits
-        }));
-
-        app.handle_key(KeyCode::Char('q'));
-        assert!(app.should_quit);
-    }
-
-    // --- Add screen: Installed tab ---
-
-    /// [verify tui.nav.keyboard]
-    #[test]
-    fn add_installed_space_toggles() {
-        let mut app = make_app(Screen::Add(make_add_screen(vec![make_installed_pack(
-            "web",
-            vec![
-                make_entry("axum", true, true),
-                make_entry("tower", true, true),
-            ],
-        )])));
-
-        app.handle_key(KeyCode::Char(' ')); // toggle first entry
-        assert!(!unwrap_add_screen(&app).installed.packs[0].entries[0].enabled);
-    }
-
-    /// [verify tui.nav.keyboard]
-    #[test]
-    fn add_installed_j_k_navigates() {
-        let mut app = make_app(Screen::Add(make_add_screen(vec![make_installed_pack(
-            "web",
-            vec![
-                make_entry("axum", true, true),
-                make_entry("tower", true, true),
-                make_entry("reqwest", false, false),
-            ],
-        )])));
-
-        app.handle_key(KeyCode::Char('j'));
-        assert_eq!(unwrap_add_screen(&app).installed.selected_index, 1);
-
-        app.handle_key(KeyCode::Char('k'));
-        assert_eq!(unwrap_add_screen(&app).installed.selected_index, 0);
-    }
-
-    /// [verify tui.nav.cancel]
-    #[test]
-    fn add_installed_esc_quits_without_changes() {
-        let mut app = make_app(Screen::Add(make_add_screen(vec![make_installed_pack(
-            "web",
-            vec![make_entry("axum", true, true)],
-        )])));
-
-        // Toggle a crate to create a pending change
-        app.handle_key(KeyCode::Char(' '));
-
-        // Esc quits without applying
-        app.handle_key(KeyCode::Esc);
-        assert!(app.should_quit);
-        assert!(
-            unwrap_add_screen(&app).changes.is_none(),
-            "Esc should not apply changes"
-        );
-    }
-
-    /// [verify tui.nav.exit]
-    #[test]
-    fn add_installed_enter_applies_when_changes_exist() {
-        let mut app = make_app(Screen::Add(make_add_screen(vec![make_installed_pack(
-            "web",
-            vec![
-                make_entry("axum", true, true),
-                make_entry("tower", false, true), // changed: was on, now off
-            ],
-        )])));
-
-        app.handle_key(KeyCode::Enter);
-        assert!(app.should_quit);
-        let state = unwrap_add_screen(&app);
-        assert!(state.changes.is_some(), "Enter should collect changes");
-        assert_eq!(state.changes.as_ref().unwrap().len(), 1);
-    }
-
-    /// [verify tui.nav.exit]
-    #[test]
-    fn add_installed_enter_does_nothing_when_no_changes() {
-        let mut app = make_app(Screen::Add(make_add_screen(vec![make_installed_pack(
-            "web",
-            vec![make_entry("axum", true, true)],
-        )])));
-
-        app.handle_key(KeyCode::Enter);
-        assert!(!app.should_quit, "Enter with no changes should not quit");
-    }
-
-    // --- Add screen: Browse tab ---
-
-    /// [verify tui.browse.search]
-    #[test]
-    fn add_browse_search_mode() {
-        let mut add_screen = make_add_screen(vec![]);
-        add_screen.tab = AddTab::Browse;
-        add_screen.browse.items = vec![make_summary("web", "1.0.0", "Web stuff")];
-        let mut app = make_app(Screen::Add(add_screen));
-
-        // '/' enters search mode
-        app.handle_key(KeyCode::Char('/'));
-        assert!(unwrap_add_screen(&app).browse.searching);
-
-        // Type search text
-        app.handle_key(KeyCode::Char('w'));
-        app.handle_key(KeyCode::Char('e'));
-        app.handle_key(KeyCode::Char('b'));
-        assert_eq!(unwrap_add_screen(&app).browse.search_input, "web");
-
-        // Backspace removes a character
-        app.handle_key(KeyCode::Backspace);
-        assert_eq!(unwrap_add_screen(&app).browse.search_input, "we");
-
-        // Esc cancels search mode
-        app.handle_key(KeyCode::Esc);
-        assert!(!unwrap_add_screen(&app).browse.searching);
-    }
-
-    /// [verify tui.nav.keyboard]
-    #[test]
-    fn add_browse_tab_switches_to_installed() {
-        let mut add_screen = make_add_screen(vec![make_installed_pack(
-            "web",
-            vec![make_entry("axum", true, true)],
-        )]);
-        add_screen.tab = AddTab::Browse;
-        add_screen.browse.items = vec![make_summary("db", "1.0.0", "")];
-        let mut app = make_app(Screen::Add(add_screen));
-
-        app.handle_key(KeyCode::Tab);
-        assert_eq!(unwrap_add_screen(&app).tab, AddTab::Installed);
-    }
-
-    // --- Add screen: Browse expanded pack ---
-
-    /// [verify tui.browse.add]
-    #[test]
-    fn add_browse_expanded_confirm_moves_to_installed() {
-        let mut add_screen = make_add_screen(vec![]);
-        add_screen.tab = AddTab::Browse;
-        add_screen.browse.expanded = Some(ExpandedPack {
-            pack: make_installed_pack(
-                "web",
-                vec![
-                    make_entry("axum", true, false),
-                    make_entry("tower", true, false),
-                ],
-            ),
-            selected_index: 0,
-        });
-        let mut app = make_app(Screen::Add(add_screen));
-
-        // Enter confirms and moves to Installed tab
-        app.handle_key(KeyCode::Enter);
-        let state = unwrap_add_screen(&app);
-        assert_eq!(state.tab, AddTab::Installed);
-        assert!(state.browse.expanded.is_none());
-        // The pack should have been added to installed
-        assert_eq!(state.installed.packs.len(), 1);
-        assert_eq!(state.installed.packs[0].short_name, "web");
-    }
-
-    /// [verify tui.browse.add]
-    #[test]
-    fn add_browse_expanded_esc_cancels() {
-        let mut add_screen = make_add_screen(vec![]);
-        add_screen.tab = AddTab::Browse;
-        add_screen.browse.expanded = Some(ExpandedPack {
-            pack: make_installed_pack("web", vec![make_entry("axum", true, false)]),
-            selected_index: 0,
-        });
-        let mut app = make_app(Screen::Add(add_screen));
-
-        app.handle_key(KeyCode::Esc);
-        let state = unwrap_add_screen(&app);
-        assert!(state.browse.expanded.is_none());
-        assert!(state.installed.packs.is_empty()); // not added
-    }
-
-    /// [verify tui.browse.add]
-    #[test]
-    fn add_browse_expanded_no_selection_discards() {
-        let mut add_screen = make_add_screen(vec![]);
-        add_screen.tab = AddTab::Browse;
-        add_screen.browse.expanded = Some(ExpandedPack {
-            pack: make_installed_pack(
-                "web",
-                vec![
-                    make_entry("axum", false, false), // nothing selected
-                    make_entry("tower", false, false),
-                ],
-            ),
-            selected_index: 0,
-        });
-        let mut app = make_app(Screen::Add(add_screen));
-
-        // Enter with no selections — pack should NOT be added
-        app.handle_key(KeyCode::Enter);
-        let state = unwrap_add_screen(&app);
-        assert_eq!(state.tab, AddTab::Installed);
-        assert!(state.installed.packs.is_empty());
-    }
-
-    // ====================================================================
-    // Tier 3: Rendering tests
-    // ====================================================================
-
-    /// Helper: render into an in-memory terminal and return the buffer content
-    /// as a string (one line per row, padded with spaces).
-    fn render_to_string(width: u16, height: u16, draw: impl FnOnce(&mut Frame)) -> String {
-        use ratatui::Terminal;
-        use ratatui::backend::TestBackend;
-
-        let backend = TestBackend::new(width, height);
-        let mut terminal = Terminal::new(backend).unwrap();
-        terminal.draw(draw).unwrap();
-        terminal.backend().to_string()
-    }
-
-    fn render_add_to_string(state: &mut AddScreen, width: u16, height: u16) -> String {
-        render_to_string(width, height, |frame| render_add(frame, state))
-    }
-
-    fn render_list_to_string(state: &mut ListScreen, width: u16, height: u16) -> String {
-        render_to_string(width, height, |frame| render_list(frame, state))
-    }
-
-    /// [verify tui.main.no-project] (partial: tests empty-state message, not greyed-out styling)
-    /// When no packs are installed, the installed tab shows a message.
-    #[test]
-    fn render_no_packs_installed_message() {
-        let mut state = make_add_screen(vec![]);
-        state.tab = AddTab::Installed;
-        let output = render_add_to_string(&mut state, 60, 15);
-        assert!(
-            output.contains("No battery packs installed"),
-            "Expected 'No battery packs installed' in:\n{}",
-            output
-        );
-    }
-
-    /// [verify tui.installed.list-packs]
-    /// Pack headers show name and version.
-    #[test]
-    fn render_installed_pack_headers() {
-        let mut state = make_add_screen(vec![
-            make_installed_pack("web", vec![make_entry("axum", true, true)]),
-            make_installed_pack("db", vec![make_entry("sqlx", true, true)]),
-        ]);
-        state.tab = AddTab::Installed;
-        let output = render_add_to_string(&mut state, 60, 20);
-        assert!(
-            output.contains("web") && output.contains("1.0.0"),
-            "Expected 'web' and '1.0.0' in:\n{}",
-            output
-        );
-        assert!(output.contains("db"), "Expected 'db' in:\n{}", output);
-    }
-
-    /// [verify tui.installed.list-crates]
-    /// [verify tui.installed.show-state]
-    /// Crate entries show checkbox ([x]/[ ]), name, and version info.
-    #[test]
-    fn render_installed_crate_entries_with_checkboxes() {
-        let mut state = make_add_screen(vec![make_installed_pack(
-            "web",
-            vec![
-                make_entry("axum", true, true),
-                make_entry("tower", false, false),
-            ],
-        )]);
-        state.tab = AddTab::Installed;
-        let output = render_add_to_string(&mut state, 60, 15);
-        assert!(
-            output.contains("[x] axum"),
-            "Expected '[x] axum' in:\n{}",
-            output
-        );
-        assert!(
-            output.contains("[ ] tower"),
-            "Expected '[ ] tower' in:\n{}",
-            output
-        );
-        // Version info should be present
-        assert!(
-            output.contains("(0.1.0)"),
-            "Expected version info '(0.1.0)' in:\n{}",
-            output
-        );
-    }
-
-    /// [verify tui.browse.list]
-    /// Browse list shows name, version, and description.
-    #[test]
-    fn render_browse_list_shows_name_version_description() {
-        let mut list_state = ListState::default();
-        list_state.select(Some(0));
-        let mut state = ListScreen {
-            items: vec![
-                make_summary("web", "2.3.0", "Web framework essentials"),
-                make_summary("db", "1.5.0", "Database toolkit"),
-            ],
-            list_state,
-            filter: None,
-        };
-        let output = render_list_to_string(&mut state, 80, 15);
-        // First pack
-        assert!(output.contains("web"), "Expected 'web' in:\n{}", output);
-        assert!(output.contains("2.3.0"), "Expected '2.3.0' in:\n{}", output);
-        assert!(
-            output.contains("Web framework essentials"),
-            "Expected description in:\n{}",
-            output
-        );
-        // Second pack
-        assert!(output.contains("db"), "Expected 'db' in:\n{}", output);
-        assert!(output.contains("1.5.0"), "Expected '1.5.0' in:\n{}", output);
-        assert!(
-            output.contains("Database toolkit"),
-            "Expected description in:\n{}",
-            output
-        );
-    }
-
-    /// [verify tui.nav.keyboard]
-    #[test]
-    fn add_browse_expanded_space_toggles() {
-        let mut add_screen = make_add_screen(vec![]);
-        add_screen.tab = AddTab::Browse;
-        add_screen.browse.expanded = Some(ExpandedPack {
-            pack: make_installed_pack(
-                "web",
-                vec![
-                    make_entry("axum", true, false),
-                    make_entry("tower", false, false),
-                ],
-            ),
-            selected_index: 1, // tower
-        });
-        let mut app = make_app(Screen::Add(add_screen));
-
-        app.handle_key(KeyCode::Char(' '));
-        let expanded = unwrap_add_screen(&app).browse.expanded.as_ref().unwrap();
-        assert!(expanded.pack.entries[1].enabled); // tower toggled on
-    }
-
-    // ====================================================================
-    // Coverage completion: main, network, new rules
-    // ====================================================================
-
-    fn render_app_to_string(app: &mut App, width: u16, height: u16) -> String {
-        render_to_string(width, height, |frame| app.render(frame))
-    }
-
-    /// [verify tui.main.sections]
-    /// The Add screen tab bar renders both "Installed" and "Browse" sections.
-    /// Note: the spec lists three sections (Installed, Browse, New project),
-    /// but New project is accessed from the Browse detail view rather than
-    /// as a top-level tab.
-    #[test]
-    fn render_add_screen_shows_section_tabs() {
-        let mut state = make_add_screen(vec![make_installed_pack(
-            "web",
-            vec![make_entry("axum", true, true)],
-        )]);
-        let output = render_add_to_string(&mut state, 60, 15);
-        assert!(
-            output.contains("Installed"),
-            "Expected 'Installed' tab in:\n{}",
-            output
-        );
-        assert!(
-            output.contains("Browse"),
-            "Expected 'Browse' tab in:\n{}",
-            output
-        );
-    }
-
-    /// [verify tui.main.always-available]
-    /// The App can be constructed and rendered without any I/O. The
-    /// constructors create a Loading screen (pure state), and rendering
-    /// that screen produces a loading message — no filesystem or network
-    /// access required.
-    #[test]
-    fn app_renders_loading_screen_without_io() {
-        let mut app = make_app(Screen::Loading(LoadingState {
-            message: "Loading battery packs...".to_string(),
-            target: LoadingTarget::List { filter: None },
-        }));
-        let output = render_app_to_string(&mut app, 60, 10);
-        assert!(
-            output.contains("Loading battery packs..."),
-            "Expected loading message in:\n{}",
-            output
-        );
-    }
-
-    /// [verify tui.main.context-detection]
-    /// Context detection (finding Cargo.toml, walking up to workspace root)
-    /// is delegated to cargo_metadata inside process_loading(). The TUI does
-    /// not implement its own project discovery — it calls load_installed_packs()
-    /// which invokes cargo_metadata on the current directory. Testing this
-    /// would be testing cargo_metadata's behavior, not our code.
-    #[test]
-    fn context_detection_delegated_to_cargo_metadata() {
-        // Intentionally empty — see doc comment.
-    }
-
-    /// [verify tui.network.non-blocking]
-    /// Non-blocking network behavior is an architectural property of the
-    /// event loop: process_loading() runs at the top of each iteration,
-    /// and event::poll() uses a 100ms timeout so the UI stays responsive.
-    /// This is a design invariant, not something expressible as a unit test.
-    #[test]
-    fn network_non_blocking_is_architectural() {
-        // Intentionally empty — see doc comment.
-    }
-
-    /// [verify tui.network.error]
-    /// Error screen renders with error message and key hints.
-    #[test]
-    fn error_screen_renders_message() {
-        let mut app = make_app(Screen::Error(ErrorScreen {
-            message: "connection refused".to_string(),
-            retry_target: LoadingTarget::List { filter: None },
-        }));
-        let output = render_app_to_string(&mut app, 60, 10);
-        expect![[r#"
-            "                                                            "
-            "                                                            "
-            "                                                            "
-            "                            Error                           "
-            "                                                            "
-            "                     connection refused                     "
-            "                                                            "
-            "         Press Enter or r to retry, Esc or q to quit        "
-            "                                                            "
-            "                                                            "
-        "#]]
-        .assert_eq(&output);
-    }
-
-    /// [verify tui.network.error]
-    /// Enter key retries by transitioning back to Loading screen with
-    /// the original target (including filter) preserved.
-    #[test]
-    fn error_screen_enter_retries() {
-        let mut app = make_app(Screen::Error(ErrorScreen {
-            message: "timeout".to_string(),
-            retry_target: LoadingTarget::List {
-                filter: Some("test".to_string()),
-            },
-        }));
-
-        app.handle_key(KeyCode::Enter);
-
-        let Screen::Loading(LoadingState {
-            target: LoadingTarget::List { filter },
-            ..
-        }) = &app.screen
-        else {
-            panic!("expected Screen::Loading(LoadingTarget::List)");
-        };
-        assert_eq!(filter.as_deref(), Some("test"));
-    }
-
-    /// [verify tui.network.error]
-    /// 'r' key also retries.
-    #[test]
-    fn error_screen_r_retries() {
-        let mut app = make_app(Screen::Error(ErrorScreen {
-            message: "timeout".to_string(),
-            retry_target: LoadingTarget::List { filter: None },
-        }));
-
-        app.handle_key(KeyCode::Char('r'));
-        assert!(matches!(app.screen, Screen::Loading(_)));
-    }
-
-    /// [verify tui.network.error]
-    /// Esc quits from error screen.
-    #[test]
-    fn error_screen_esc_quits() {
-        let mut app = make_app(Screen::Error(ErrorScreen {
-            message: "error".to_string(),
-            retry_target: LoadingTarget::Add,
-        }));
-
-        app.handle_key(KeyCode::Esc);
-        assert!(app.should_quit);
-    }
-
-    /// [verify tui.network.error]
-    /// 'q' quits from error screen.
-    #[test]
-    fn error_screen_q_quits() {
-        let mut app = make_app(Screen::Error(ErrorScreen {
-            message: "error".to_string(),
-            retry_target: LoadingTarget::Add,
-        }));
-
-        app.handle_key(KeyCode::Char('q'));
-        assert!(app.should_quit);
-    }
-
-    /// [verify tui.new.template-list]
-    /// Template listing is shown in the DetailScreen's selectable items
-    /// (already tested by detail_selectable_items_includes_all_sections).
-    /// The "from crates.io" aspect requires a network fetch in
-    /// process_loading() which is not unit-testable without mocking.
-    #[test]
-    fn template_list_covered_by_detail_screen_tests() {
-        // Intentionally empty — see doc comment.
-        // Real coverage: detail_selectable_items_includes_all_sections.
-    }
-
-    /// [verify tui.new.create]
-    /// Project creation shells out to `cargo bp new` via
-    /// PendingAction::NewProject in execute_action(). This spawns an
-    /// external process (`cargo bp new`), which is not unit-testable.
-    #[test]
-    fn new_project_creates_via_external_process() {
-        // Intentionally empty — see doc comment.
-    }
-
-    // --- Preview screen ---
-
-    #[test]
-    fn preview_esc_returns_to_detail() {
-        let detail = make_detail(&["serde"], &["default"], &[]);
-        let mut app = make_app(Screen::Preview(PreviewScreen {
-            content: Text::from("test content"),
-            template_name: "default".to_string(),
-            scroll: 0,
-            line_count: 1,
-            detail: Rc::new(detail),
-            selected_index: 2,
-            came_from_list: true,
-        }));
-
-        app.handle_key(KeyCode::Esc);
-        assert!(matches!(app.screen, Screen::Detail(_)));
-        if let Screen::Detail(state) = &app.screen {
-            assert_eq!(state.selected_index, 2);
-            assert!(state.came_from_list);
-        }
-    }
-
-    #[test]
-    fn preview_scroll_down_and_up() {
-        let detail = make_detail(&[], &["default"], &[]);
-        let mut app = make_app(Screen::Preview(PreviewScreen {
-            content: Text::from("line1\nline2\nline3\nline4\nline5"),
-            template_name: "default".to_string(),
-            scroll: 0,
-            line_count: 5,
-            detail: Rc::new(detail),
-            selected_index: 0,
-            came_from_list: false,
-        }));
-
-        app.handle_key(KeyCode::Down);
-        if let Screen::Preview(state) = &app.screen {
-            assert_eq!(state.scroll, 1);
-        }
-
-        app.handle_key(KeyCode::Char('j'));
-        if let Screen::Preview(state) = &app.screen {
-            assert_eq!(state.scroll, 2);
-        }
-
-        app.handle_key(KeyCode::Up);
-        if let Screen::Preview(state) = &app.screen {
-            assert_eq!(state.scroll, 1);
-        }
-
-        app.handle_key(KeyCode::Char('k'));
-        if let Screen::Preview(state) = &app.screen {
-            assert_eq!(state.scroll, 0);
-        }
-    }
-
-    #[test]
-    fn preview_scroll_clamps_at_bounds() {
-        let detail = make_detail(&[], &["default"], &[]);
-        let mut app = make_app(Screen::Preview(PreviewScreen {
-            content: Text::from("line1\nline2"),
-            template_name: "default".to_string(),
-            scroll: 0,
-            line_count: 2,
-            detail: Rc::new(detail),
-            selected_index: 0,
-            came_from_list: false,
-        }));
-
-        // Scroll up at 0 stays at 0
-        app.handle_key(KeyCode::Up);
-        if let Screen::Preview(state) = &app.screen {
-            assert_eq!(state.scroll, 0);
-        }
-    }
 }
